move some models to shared models
